requirements.txt
APScheduler==3.11.2
Blinker==1.9.0
certifi==2026.1.4
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
Flask==2.3.3
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pytz==2025.2
requests==2.32.5
six==1.17.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.5
watchdog==6.0.0
prometheus-client==0.20.0
pyyaml>=6.0
jsonpath-ng>=1.6.0
gunicorn>=21.0.0
gevent>=23.0.0